Rename `utils` module to `util`.
