+++ lib/entry.js
@@ -3,7 +3,7 @@
 var _ = require('lodash'),
     os = require('os'),
     Alias = require('./alias.js'),
-    utils = require('./utils.js');
+    util = require('./util.js');
 
 /*----------------------------------------------------------------------------*/
 
@@ -89,7 +89,7 @@ function getAliases(index) {
     if (result) {
       result = result.replace(/(?:^|\n)[\t ]*\*[\t ]*/, ' ').trim();
       result = result.split(/,\s*/);
-      result.sort(utils.compareNatural);
+      result.sort(util.compareNatural);
 
       result = _.map(result, function(value) {
         return new Alias(value, this);
@@ -327,7 +327,7 @@ function getMembers(index) {
     var result = getValue(this.entry, 'member');
     if (result) {
       result = result.split(/,\s*/);
-      result.sort(utils.compareNatural);
+      result.sort(util.compareNatural);
     }
     this._members = result || [];
   }
