+++ lib/generator.js
@@ -3,7 +3,7 @@
 var _ = require('lodash'),
     Entry = require('./entry.js'),
     getEntries = Entry.getEntries,
-    utils = require('./utils.js');
+    util = require('./util.js');
 
 var push = Array.prototype.push,
     specialCategories = ['Methods', 'Properties'],
@@ -246,11 +246,11 @@ function generateDoc(source, options) {
     _.pull.apply(_, [tocGroups].concat(specialCategories));
 
     // Sort categories and append special categories.
-    tocGroups.sort(utils.compareNatural);
+    tocGroups.sort(util.compareNatural);
     push.apply(tocGroups, specialCategories);
   }
   else {
-    tocGroups.sort(utils.compareNatural);
+    tocGroups.sort(util.compareNatural);
   }
   // Start markdown for TOC categories.
   var tocMarkdown = ['<!-- div class="toc-container" -->\n'];
@@ -266,8 +266,8 @@ function generateDoc(source, options) {
       organized[group].sort(function(value, other) {
         var valMember = value.getMembers(0),
             othMember = other.getMembers(0);
-  
-        return utils.compareNatural(
+
+        return util.compareNatural(
           (valMember ? (valMember + getSeparator(value)) : '') + value.getName(),
           (othMember ? (othMember + getSeparator(other)) : '') + other.getName()
         );
